covHisto_WG needs to be long[] to not overflow
When computing a histogram across the whole genome, you'll need to use an array of longs to avoid integer overflow (so you don't get a histogram that says that -1414732325 bases were at zero coverage, for instance).

I notice that the basesWithOneHitorMore_WG variable and others like it were changed to longs instead of ints, so that's good. On the other hand, it might make more sense to just compute those from the histogram instead of keeping individual tallies as you go along.